k1c: Enable CONFIG_BLK_DEV_INITRD
Summary:
This option is required to use a initramfs. It is easier to add this
option alone in the default config instead of modifying the .config when
building the initramfs because enabling this option requires other
options. So instead of adding them one by one it makes more sense to add
it in the default config.

Test Plan: Boot until VFS panic

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D117