k1c: Implement syscall entry
Summary:
This commit implements the syscall entry assembly

A documentation is included which described the coarse grained steps. In  assembly, there is much more comments to clearly explain what is done during the call.

Ref T5691

Test Plan: Wait for user space to be ready, run the code and check syscalls are working

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Maniphest Tasks: T5691

Differential Revision: https://phab.kalray.eu/D89