k1c: Remove LTLB[1] during paging init
Summary:
The first entry of LTLB contains the mapping of 512Mo of SMEM. This was
used during the early boot before the jump to virtual memory space. Now
we are in virtual memory space so this mapping is not needed anymore and
we can remove it.

Test Plan: Boot until VFS panic

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D120