k1c: Add documentation and info about MMU management
Summary:
This patch updates documentation about how physical memory is
managed by Linux and on the MPPA. It also details what are the
possibilites of levels in page table according to the page size. Finally
it adds some message during the boot to help the understanding of the
memory allocation.

Test Plan: Boot until panic on pgd_alloc not implemented

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D121