k1c: First implementation of pgd_alloc and do_page_fault
Summary:
This patch does the minimal to be able to create the memory space for
the init process. In this first implementation we are using a 3 level
page table with a 4KB page. The user memory space is 512GB.
To achieve this this patch implements:
  - pte_alloc_one(), __pte_free_tlb()
  - pgd_free(), pgd_alloc()
  - pmd_free(), pmd_alloc()
  - pud_none(), pud_populate(), set_pud()
It also does some cleanup of the swapper_pg_dir and fixes the value of
pud returns when calling pud_page_vaddr(). Finally it fills the page
table with PGD, PMD and PTE entry to be able to reach the do_page_fault()
handler.

Test Plan: Boot until: "Kernel panic - not syncing: do_page_fault: not yet implemented"

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D123