k1c: add a panic handler to emit syscall
Summary:
In order to keep validation point in Linux, add a panic handler with an
expected panic string. If the panic message matches the expected string,
exit with 0. If not exit with 1.
Moreover, set the current expected message to pgd_alloc not implemented
yet. \n is already embedded into panic printed string so remove it from
pgd_alloc panic message.

Test Plan:
Integration should be passed and once finish, remove the grep
from build.rb to see it sill works.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Differential Revision: https://phab.kalray.eu/D127