k1c: Implement basic ioremap
Summary:
In order to be able to use ioremap, add a basic implementation.
This implementation uses the peripheral memory mapping which is added in
the ltlb. Later, we will need to handle it more dynamically but since we
don't have page walking yet, use this static mapping.

Depends On D128

Test Plan: Call ioremap from a driver and check it works.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Differential Revision: https://phab.kalray.eu/D129