k1c: Implement cmpxchg
Summary:
cmpxchg is based on acswap(d|w) whic atomically swaps a value if the
previous one matches the one provided by the user. On k1c, the acswap is
a boolean compare exchange meaning that the old value is not returned
but only a boolean which is 1 on success, and 0 on failure. Since
cmpxchg needs to return the old value, we must load it. In case of
success, we simply return the old value provided by the user since by
construction this was the value when swapping occurs. In case of error,
we have to distinguish 2 cases. The first one is that if after the load,
the value is different from the one provided by the user, we can
return it directly, the user will comapre it with its value and see it
failed. However, if we load the value and find that it is equal to the
one provided by the user (which might happen if someone exchanges the
value before our turn), then we necessary have to retry until we
failed AND the loaded value after failure is different.

Ref T5334

Test Plan: Check that Linux still validates

Reviewers: O51 linux!

Maniphest Tasks: T5334

Differential Revision: https://phab.kalray.eu/D136