k1c: Implement pmd and improve documentation
Summary:
This patch implements the pmd level and improves the debug by
enabling the kernel debug message. It also enables the output of
memblock information and removes the output of the function
raw_copy_from_user(). Finally it also documents PFN and
most of macro/functions used for the manipulatation of the
page table.

Test Plan: Boot until update_mmu_cache is not yet implemented

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D144