k1c: Fix a bug in the computation of the PTE offset
Summary:
This patch fixes a bug when computing the PTE offset. Without the cast
to (pte_t *) the addition is not considered as the addition of pte_t and
so the result was not the index in the page but directly the value of
the index. We also made some cleanup.

Test Plan: boot until update_mmu_cache is not yet implemented

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D152