k1c: add missing scratchpad setup in assembly
Summary:
When calling C function from assembly, they expect a potential
scratchpad area. If not set, the stack can be overwritten. Add this
missing scratchpad and also remove unneeded indirect call. We can handle
it with the current text size and symbol location with an immediate.

Test Plan: Nothing really particular since we never triggered a bug realted to this.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Differential Revision: https://phab.kalray.eu/D150