k1c: Start the implementation of the do_page_fault
Summary:
This patch implements the case where a PTE entry is present but the
mapping is not present in the JTLB. We use A-Modified for the entry
status to avoid to catch a trap WRITETOCLEAN. The trap WRITETOCLEAN is
used to implement the copy-on-write and we will have to implement it at
some point. Currently we just prevent to catch ATOMICTOCLEAN and
WRITETOCLEAN traps to reach the return to user from kernel thread. Note
that this patch also wraps the do_page_fault() in k1c_trap_nomapping() to
have something coherant with all other traps like WRITETOCLEAN that we will
immplement.

Test Plan: Boot until starting init process that will shutdown the
machine properly.

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D170