k1c: exit with 1 instead of 0 on halt
Summary:
When taking a panic before our panic syscall handler has been
registered, it will end on a machine halt. However, currently, our
machine halt returns with 0. In case of error, we will think everything
is ok (which is not !). Fix that by always returning 1 in machine halt.
If panic happens too early, we will exit with an error, which is what we
want since it should not happen.

Test Plan: Run integration and validates that we run up to the expected point and can never crash early again.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D171