k1c: Remove -ffarcall flag from build
Summary:
farccall is needed if we now we are going to do far calls. currently,
this is not the case and it can degrade performacnes by always doing
indirect calls (ie, set a register and "call" it.

Test Plan: Check that everything still compiles and run up to the current point.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D186