k1c: Fix empty syscall table
Summary:
Due to missing unistd.h in asm, syscall table was empty during
generation. In order to fix that, we add an unistd.h which also define
that the arch want sys_clone. This header must not have incldue guard
since it will be included multiple times !
When instantiating all the syscalls, some of them were missing (mmap and
rt_sigreturn). Stub them with panic to compile.

Ref T6203

Test Plan: Integrate and check that the integration is successfull

Reviewers: O51 linux!

Maniphest Tasks: T6203

Differential Revision: https://phab.kalray.eu/D195