k1c: Add CONFIG_DEBUG_EXCEPTION_STACK to debug stack
Summary:
In order to be sure stack was not messed up after leaving kernel, add a
check in exception handlers.
This mecanism is fairly simple, we push the stack pointer on the stack
when entering and we will check it on exit.

Depends On D200

Test Plan: Voluntarly mess with $sp during exception handling to see the checker is triggered

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Differential Revision: https://phab.kalray.eu/D201