k1c: Check write permission using VMA flags
Summary:
This patch modifies the check of the write permission because pgprot_val
cannot be compared to _PAGE_WRITE and it seems more obvisous to use the
flags of VMA directly instead of making a conversion that is not very
understandable for now.

Test Plan: Be able to run the new initramfs. Data area is now writeable.

Reviewers: O51 linux!, cleger

Differential Revision: https://phab.kalray.eu/D231