k1c: Add spc, sps and cs save in syscall
Summary:
We were reenabling interrupts in syscall but not clearing ET, this was a
first problem lading to core error since we can not take an exception
inside an exception if ET is not cleared. This is now done by the wfxl
with the good value.
Once exception were ok, the second problem was that we did not save
enough context to be able to take an exception during sycall. This is
now possible due to sps, spc being saved additionnally as cs and ra.
Finally, we are now able to take an exception inside a syscall.

Test Plan: force interrupt to happen inside syscall by calling syscall in a tight loop.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D238