k1c: Fix wrong cache policy for mmu entries
Summary:
Previous cache policy was Device for data and Uncached for instruction.
This is not what we want. We want to be writethrough for data end cached
for instructions. Fix that by using the good policy on refill.

Test Plan: Run image in integration and check everything is still correct

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Differential Revision: https://phab.kalray.eu/D242