k1c: Add APIC mailbox driver
Summary:
In order to handle SMP, add APIC mailbox driver. This controller is an
msi controller with IPI capability. IPI are handled using a sperate
domain. This domain is built on top of the GIC one using dedicated
lines.

Depends on D305

Test Plan: Run a test with ITGEN (uart for instance)

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D386