k1c: Add needed header for headers_install to work
Summary:
In order to build uClibc, the linux headers are needed. If some of them
are missing, it fails. Add the needed one as generic to allow
headers_install to work.
Move sigcontext.h to uapi/asm since this were it is expected by uClibc.
Moreover, fix unistd to be used by uClibc

Test Plan: Install headers and build uClibc

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Differential Revision: https://phab.kalray.eu/D246