k1c: Handle properly no mapping trap for user access
Summary:
When accessing user buffer, we might trigger a pagefault. When handling
this pagefault, we were only checking fixup if page_fault was disabled
or if there was no mm. However, there is another path which is after
bad_area. Implement basic handling to handle such case.

Test Plan: Check that initramfs now works with this patch

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D284