k1c: Add the retry path in do_page_fault
Summary:
The retry is needed for example in the COW. If a page fault occurs on a
read only page that is marked as COW, then a new page is created and we
need to retry to load it. It is also important because this patch adds
the management of the mmap->sem.

Test Plan: init must run successfully

Reviewers: O51 linux, cleger

Reviewed By: O51 linux, cleger

Differential Revision: https://phab.kalray.eu/D218