k1c: handle _PAGE_DEVICE in tlb_refill routine
Summary:
In order to handle proper ioremap, we need a way to install tlb entries
using the proper device policy. The policy can be found using the
pte_entry bit _PAGE_DEVICE which will be set when using ioremap.
The refill then checks this flag and set the cache policy in the tlb
entry.

Ref T6955

Test Plan: iWait for ioremap to be ready and check device access to be ok.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Maniphest Tasks: T6955

Differential Revision: https://phab.kalray.eu/D298