k1c: implement the management of the ASN
Summary: This patch implements the management of ASN for user space
process. There is cache for each CPU that is used to keep track of
current ASN. When a new process is created its ASN is set to 0 and its
running CPU is set to -1. When scheduled it gets a new ASN and CPU is
set. As the ASN has 9-bit, each 512 process we cycle. We are using the
higher bits (10 to 63) to check if the ASN used by a process has been
provided from the current cycle or if it is obsolete and needs to be
created again. Pages that belongs to kernel space (bit 40 is set to 1)
have the bit GLOBAL set to 1. For user pages it is set to 0 and ASN is
used. All pages with GLOBAL bit set to 1 must have rights set to NA when
PM == 0. Otherwise it is a bug and security issue.

Test Plan: Fork process in the init and check that ASN is well updated

Reviewers: O51 linux!, cleger

Maniphest Tasks: T7463

Differential Revision: https://phab.kalray.eu/D383