k1c: Update panic check for init exit
Summary:
The user space now goes up to the proper exit call (exit(0)).
When exiting, the kernel outputs a message of panic since init exited
with status 0. This is correct for us and we want to enforce that for
the moment.

Test Plan: Run the kernel with the latest init to chekc it goes up to the exit

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D399