k1c: Fix wrong naming for PANIC_SCALL and use strncmp
Summary:
In the config, it was referenced as PANIC_SYSCALL but int he config, it
was PANIC_SCALL. This has never been compiled in but it worked for some
mystic reason. It will be squashed in original commit.
Moreover, since \n are not handled correctly in kconfig, use strncmp to
shorten the comparison length.

Test Plan: Run with the latest string containing an \n

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D400