k1c: Add DAME handler
Summary:
In case of a Data Asynchronous memory error, an interrupt will be
triggered to signal the error. Unfortunately, this error can't be
handled and we can simply panic to stop the kernel.
The only way to debug this is to disable streaming bu setting $ps.use =
0.

Test Plan: trigger a data asynchronous memory error from kernel space.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Differential Revision: https://phab.kalray.eu/D390