k1c: Fix wrong setup of cache policy
Summary:
Commit 25f4615d2 added support for proper permissions table but by doing
so, $r3 was cloberred. Unfortunately, since we don't have any devices
yet in the kernel, it worked correctly since $r3 was always reassigned
by the conditionnal move. Using a device with ioremap (yet to be integrated)
showed that the access permissions where wrong (due to $r3 being cloberred)
leading to protection traps.

Test Plan: Enable a peripheral and be sure it correctly access its registers.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D407