k1c: Safer start sequence for mmu and exception
Summary:
When booting, we clear et early to be sure we can emit syscall to jtag
and avoid taking double traps. In order to be safe, add an "early"
exception vector which will catch any trap happening while the kernel is
not ready to handle them. Moreover, remove the smem mapping quite early
to catch any null call problem. This has the side effect of removing the
local peripheral mapping which is anyway not used anymore since ioremap
has been implemented.

Depends on D409

Test Plan:
Run init and tigger a trap early ($ps.et = 1) to check it correctly go to the
exception handler

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D410