k1c: Add section for custom rm firmware
Summary:
In order to run SMP kernel, we need the L2$ controller to be running on
RM. To do this simply, we add a section for rm_firmware code. By
default, this section contains the default rm_power_off code. When L2$
controller will be needed, we will simply update the content of
.rm_firmware section using objcopy --update-section. The firmware will
be built out of linux sources to keep a clean source tree and will be
then patched into the kernel binary. When patched, the section will then
contains the L2$ code and run it without any modification.

Test Plan: Wait for SMP to be ready and integrate L2$ controller.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Differential Revision: https://phab.kalray.eu/D424