k1c: save real stack pointer in pt_regs
Summary:
When entering the kernel, the original stack pointer was saved in
TASK_THREAD_USER_SP. However, this was complicated to restore it and to
do ret_from_fork cleanly since some work was needed to gather free
registers and recopy other thing.
With this new mecanism, the original stack pointer is saved in a
temporary register before being saved in PT_Q12. This allows to simply
restore $r12 at the end of the context switch without complicated
comparisons. For ret_from_fork, this is now also simpler since we do not
need to recopy $r12 in TASK_THREAD_USER_SP and we can simply leave it in
r12 which will be restored.

Ref T7335

Test Plan:
Run the kernel in integration to check we still execute correctly.
Moreover, check that the user stack is correctly restored in traces.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Maniphest Tasks: T7335

Differential Revision: https://phab.kalray.eu/D438