k1c: Fix issues with ASN
Summary:
This patch fixes the invalidation of an entry in the TLB. Instead of
flushing the whole set it only invalidate the entry if the address and
the ASN find are matching the ones passed as parameters.

It also activates the new context when a flush of the mm occurs. Now
flushing the mm corresponds to generate a new context (like before) but
also activate it. Otherwise it is a bug and the old entries are still
valid until a context switch occurs and activate the new context.

Finally we move the validation of the ASN into one place if
CONFIG_K1C_DEBUG_ASN is selected instead of defining a different
function if it is defined or not.

Test Plan:
Improve initramfs test with memory allocation and flush of TLB when
the memory is released.

Reviewers: cleger, O51 linux!

Differential Revision: https://phab.kalray.eu/D446