k1c: Add a macro to set a value for a given field of an SFR
Summary:
This patch adds a macro to set a value for a given field of an SFR. The
idea is to have the name of the SFR as parameter to tell you that it
will be written. This is a convention that we take: if the SFR is passed
as a parameter it means that there is an access to this SFR. For example
before this patch a call to k1c_mmu_select_set(0) made an access to MMC
but you are not aware without reading the function. It can lead if you
don't pay enough attention to perform several consecutive accesses to an SFR
without knowning it. This convention try to enphasis the fact that an
access to SFR is done.

Test Plan: Boot as before

Reviewers: cleger, O51 linux!

Differential Revision: https://phab.kalray.eu/D451