k1c: debug TLB access
Summary:
This patch adds an option to configure the size of the ring buffer used to
store accesses done into the TLB. This option is available if K1C_DEBUG
is set. In the table we store TEL, TEH and also the SS, SB, SW, ASN and
the type of access (read, write or probe). It also add a command in gdb
to dump the last ten operations that occured in the TLB. The syntax is:
lx-k1c-tlb-access <X>, where X is the CPU or 0 by default.

Test Plan: Run vmlinux in gdb and check that table k1c_tlb_access is updated

Reviewers: cleger, O51 linux!

Differential Revision: https://phab.kalray.eu/D457