k1c: handle DAME securely.
Summary:
In order to handle DAME interrupt securely, we need to guarantee that
the user triggered DAME interrupts will not be serviced by the kernel in
kernel context. This implies a few barrier and some checking before
returning to user.

Ref T7939

Test Plan:
Manually tested by setting $ilr during debugging and checking that
we match the expected effects.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: pgmassas, vray, gthouvenin

Maniphest Tasks: T7939

Differential Revision: https://phab.kalray.eu/D441