k1c: Add IPI controller
Summary:
Add the IPI controller support for Coolidge. This IPI controller driver is
rather simple and is intended to be used by the SMP layer. The k1c_init_ipi
function will be called in smp_init to verify it works.

Ref T7598

Test Plan: Run it in smp to verify it works

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: gthouvenin

Maniphest Tasks: T7598

Differential Revision: https://phab.kalray.eu/D359