k1c: Fix wrong pt_regs passed to irq/trap handlers
Summary:
On k1c defconfig, CONFIG_DEBUG_EXCEPTION_STACK is enabled. This featrue
was implemented to allow stack pointer checking on exception entry/exit.
However, when this feature was added, I did not check that the pt_regs
address passed to trap_handler and do_irq was ok. When the stack is
checked, we add 8 bytes to the stack in order to stock the entry stack
pointer. Unfortunately, this $sp is passed to do_IRQ and trap_handler,
leading to a misplaced pt_regs ponter. Fix that by adding a parameter
to save_regs_for_exception which will copy the saved pt_regs pointer
into that register.

Test Plan: Check that pt_regs pointer is matching stack address in trace.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D463