k1c: remove NULL trapping mecanism
Summary:
This mecanism was prepared to handle NULL pointer dereferencement.
However, this was not a good solution for the following point:
- User can trap on a lot of other addresses.
- To use prefetch, GCC must need them it to be non-trapping
In order to have a functional prefetch on both kernel and user, disable
mmc.sne and mmc.spe to be sure dtouch are not trapping at all.

Ref T8341

Test Plan:
Verify that kernel dtouchl at 0x0 still works and that user can do dtouchl at
invalid addresses withotu dying

Reviewers: O51 linux!

Subscribers: mpoulhies, gthouvenin

Maniphest Tasks: T8341

Differential Revision: https://phab.kalray.eu/D474