k1c: Add spc in user context
Summary:
In order to handle user registers restoration properly, add $spc in
user_pt_regs. When executing a signal handler, we set spc to the value of the
signal handler function. However, we did not save the previous $spc in user
context. Consequently, when returning from signal handler, the user spc was
not restored, leading to invalid program counter. Now that spc is saved, the
correct spc is restored to return to userspace pc before siganl handling.

Test Plan: Run init with signal testing.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Differential Revision: https://phab.kalray.eu/D499