k1c: call do_page_fault() on protection error
Summary:
When a process is cloned, PTE are coped from the father to the child. To
allow COW the write bit is removed from PTE. During the copy the PTE of
the child is added in the TLB with write disabled. On
the first write access the protection trap is yield and we call
do_page_fault(). In the common path of the do_page_fault(), if a COW is
detected (PTE is not writable but the VMA associated to it is writeable) a
new physical page is allocated (it is handled via do_wp_page()) and the PTE is
updated accordingly with write permission. k1c_trap_nomapping() has been
removed because it was just a call to do_page_fault(). To be coherent we
also renamed k1c_trap_writetoclean() into do_writetoclean().

Test Plan: Add a test that forks in the initramfs and check that everything is
working like a charm.

Reviewers: cleger, O51 linux!

Maniphest Tasks: T8399

Differential Revision: https://phab.kalray.eu/D502