k1c: Save and restore user stack from kernel stack
Summary:
do_signal can modify the user stack pointer to create a frame for
rt_sigreturn in order to carry information. Before that, the stack was
restored from TASK_THREAD_USER_SP which is restored as-is before
returning to userspace leading to an incorrect stack pointer for
rt_sigreturn.

Ref T6696

Test Plan:
Run init with signals testing to be sure the stack has been modified when
returning to signal handler.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Maniphest Tasks: T6696

Differential Revision: https://phab.kalray.eu/D497