k1c: set FIRST_USER_ADDRESS to 0
Summary:
When a process exits free_pgtables() is called to release the page
table. This function is calling free_pgd_range(). For small addresses,
for example for a VMA that starts at 0x10000 and ends at 0x1b000,
the addr &= PMD_MASK is returning 0. If addr is strictly smaller
than FIRST_USER_ADDRESS then PMD_SIZE is added. PMD_SIZE is 1 << 21 in
our case (0x200000). Later in the code there is a comparaison between
this new computed address (0x200000) and the end of the VMA zone that is
in this case 0x1b000. As the new computed address is greater than the
end of the VMA free_pgd_range() returns without freeing the pages of the
page table. To fix the issue we set FIRST_USER_ADDRESS to 0.

Ref T8505

Test Plan: Check that BUG: non-zero pgtables_bytes on freeing mm: 8192 is no longer present in the log

Reviewers: O51 linux!

Maniphest Tasks: T8505

Differential Revision: https://phab.kalray.eu/D521