k1c: Implement jump label support
Summary:
Implement jump_label support based on goto and nop.
For this purpose, goto is generated and the addressing range is check to
ensure there will be no problems. Moreover, also add SMP support by
using the same mecanism as arm64 and xtensa.

Test Plan: jump  label testsuite is enabled in debug config to test it.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D545