k1c: fix wrong and dangerous stack restoration
Summary:
When the stack restoration was modified, the original r12 was restored
using the load from PT_R12. However, there was also the ipt_regs space
reclaiming which is absolutely useless since $sp is also overwritten.
The result is a dangerous bundle with $r12 assigned twice.

Test Plan: Run integration and vmlinux validation to check it still works correctly.

Reviewers: O51 linux!

Differential Revision: https://phab.kalray.eu/D559