k1c: Flush tlb_page when entering update_mmu_cache
Summary:
update_mmu_cache can be called to update an existing entry. If so, we
need to delete any existing entries from the TLB since it will result in
multiple mapping for the same v@ which result in an undefined behavior.
This is even more useful in SMP since some TLB entries might be stalled
on other CPUs.

Ref T8831

Test Plan: Run with SMP mode which trigger multimappings.

Reviewers: O51 linux, pkerbrat

Reviewed By: O51 linux, pkerbrat

Subscribers: alfred

Maniphest Tasks: T8831

Differential Revision: https://phab.kalray.eu/D581