k1c: Fix wfxl sps not alone in the bundle
Summary:
The _restore_regs macro from entry.S bundles wfxl $sps
with a load. This is forbidden.

Test Plan:
This is tested when debugging a linux application. The
wfxl $sps will try to enable SME which will trap to ring level 0,
the owner of SME field when k1-jtag-runner is used. k1-jtag-runner
will apply the wfxl and will skip the bundle.

Reviewers: O51 linux, cleger

Reviewed By: O51 linux, cleger

Subscribers: alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D681