k1c: enable huge pages of 2Mo
Summary:
This patch enables huge pages of 2Mo. It adds 3 extra bits in the PTE. One
to specify if the page is huge and 2 more extra bits to specify the
size of the huge page. It is not strictly required but it helps to get
the size of the page in assembly and it will be needed when we will
introduce the huge page of 512M. Currently huge page of 512M doesn't fit
in PMD directly nor in PGD. Only huge page of 2M fits in PMD. It is the
same for 64K huge page that doesn't fit in PTE. It requires 16
contiguous entries in PTE or a rewrite of the page table.

Ref T9355

Test Plan: Run all LTP hugemmap* tests

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9346

Differential Revision: https://phab.kalray.eu/D970