k1c: Remove useless double initialization.
Summary:
JTLB cleanup, PMJ setup and error clearing is already done in head.S in
early boot. There is no need to do it twice so remove this part of code
and simply clear SMEM entry as before.
While removing this part of code, jtlb cleanup is not used anymore in
this file. Move the jtlb cleanup where it is only needed (ie in tlb.c
for local_flush_tlb_all).

Test Plan:
Checked that the cleanup is correctly done using DEBUG config to dump
TLBs.

Reviewers: O51 linux, gthouvenin

Reviewed By: O51 linux, gthouvenin

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D775