k1c: handle fault cause in do_page_fault
Summary:
When user tries to read/write/exec at an address and take a protection
trap, we must check the vma flags to allow this access. If vm_flags are
not allowing the access, we must kill the process. Access type is
extracted from $es (exception status) and handled accrodingly to
vm_flags.

Ref T9411

Depends On D780

Test Plan: Test added in coolval.

Reviewers: O51 linux!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9411

Differential Revision: https://phab.kalray.eu/D781