k1c: Add support for hardware breakpoint/watchpoint
Summary:
When a hardware watchpoint/breakpoint is owned by a more
privileged PL (e.g. the k1-jtag-runner debug routines (PL0) own them by
default) and it is required by gdbserver, ptrace requests it from the
owner. k1-jtag-runner gives the ownership to the requesting PL if the
watchpoint/breakpoint is not currently used by PL0, else ptrace informs
gdbserver that the resource is not available.
Ref T9529

Test Plan:
Tested by debugging a linux application. Waiting for a second
connection between the host and MPPA to add a test in integration.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9529

Differential Revision: https://phab.kalray.eu/D900