k1c: mask asn to remove cycle part
Summary:
When context is switch, process asn is written in $mmc. However, the
process asn also contains the "cycle" part of the ASN. This part is only
relevant to the kernel and must be masked. When not masking, some bits
of $mmc were overwritten due to high cycle.
Since a picture is worth 1000 words:

$mcc layout:
```
+-----------------+-----+--------+---+------------+
|       ...       | SNE |   Res  | S |    ASN     |
+-----------------+-----+--------+---+------------+
                    14   13    10  9  8          0
```
When process asn + cycle is equal to 0x4000, bit 14 is "1". If this
value is written to $mmc without being masked, then, bit 14 is toggled
to 1 and thus, SNE will set speculative accesses to trigger nomapping.
Then, the subsequent prefetch calls will trigger a nomapping and since
prefetch is also done for NULL pointer, the kernel will panic for NULL
pointer dereferencement.

Ref T9656

Test Plan: A tests which generate a lot of asn by spawning process has been added

Reviewers: O51 Linux Coolidge, ysionneau, gthouvenin

Reviewed By: O51 Linux Coolidge, ysionneau, gthouvenin

Subscribers: ysionneau, #linux_coolidge_cc

Maniphest Tasks: T9656

Differential Revision: https://phab.kalray.eu/D1016