k1c: Add macro for phys addr in assembly
Summary:
In order to compute physical addresses from virtual addresses (symbol
for instance), add a macro. This will be used for l2$ in order to
convert virtual address to physical (dtb for instance)

Test Plan:
Add __PA in assembly to check it correctly compiles. Moreover, this patch
is used in l2 cache driver to pass device tree to firmware.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D996