k1c: slow path for sys_clone
Summary:
Since only clone syscall needs callee savec register to be preserved
across the call, save them in a separate slow-path. This allows to keep
a nice fast-path for syscall handling.
This new slow path redefine the sys_clone in syscalls.h and implement it
in assembly. callee-saved registers are now saved only is this case.

Ref T9505

Test Plan: Checked that clone tests are still working as expected

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T9505

Differential Revision: https://phab.kalray.eu/D945