k1c: Add support for TCA registers saving
Summary:
In order to allow the user to use the TCA, we need to save its registers
between context switches to avoid cloberring other tasks register
content. This is done only when the XMF bit of $cs is set which means
that the coprocessor was used. Restoration is also only done when a
context was previously saved.
Moreover, since the coprocessor register set uses 4Ko, make this support
optionnal since it will have an heavy cost on the context switch.

Test Plan: N/A yet, WIP

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D964