k1c: Manage all huge page size
Summary:
This patch adds huge page of 64Ko and 512Mo. To support this size we use
several contiguous entries in the page table. For huge pages of 64Ko we
will need to use 16 entries in the PTE and for huge pages of 512Mo we
will use 256 entries in PMD.

Ref T9355

Test Plan: Add a test that uses pages of 64Ko in coolval

Reviewers: O51 Linux Coolidge, cleger

Subscribers: cleger, #linux_coolidge_cc

Maniphest Tasks: T9355

Differential Revision: https://phab.kalray.eu/D1184