remoteproc: Add k1c cluster remoteproc driver
Summary:
Implement the remote proc backend for k1c. This backend uses the ftu to start
remote processor and required tcm and dsu to be defined in device tree. It also
uses k1c mailboxes to communicate with remote processor.

Ref T8894

Test Plan: Wait for test to be available to run a remote processor correctly.

Reviewers: O51 Linux Coolidge, jmaselbas, vchardon

Reviewed By: O51 Linux Coolidge, vchardon

Subscribers: vchardon, #linux_coolidge_cc, pkerbrat, ysionneau, alfred

Maniphest Tasks: T8894

Differential Revision: https://phab.kalray.eu/D528