k1c: clean bits ignored in PN and FN
Summary:
This patch clean bits ignored in PN and FN according to the page size.
To do this it cleans the offset of virtual and physical addresses passed
to the function that creates TEH and TEL entries.

Ref T9849

Test Plan:
Check that logs generated by simulator that complains about bits
ignored in PN not set to 0 disappeared.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: ysionneau, alfred, cleger, #linux_coolidge_cc

Maniphest Tasks: T9849

Differential Revision: https://phab.kalray.eu/D1191